refactor `Relation` into its own file
